requirements.txt
# ============ Core ML / PyTorch ============
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
lightning==2.6.1
pytorch-lightning==2.5.5
accelerate==1.12.0
transformers==4.57.6
tokenizers==0.22.2
safetensors==0.6.2
peft==0.11.1
bitsandbytes==0.48.1
diffusers==0.36.0
datasets==2.21.0
sentencepiece==0.2.1

# ============ Audio Processing ============
librosa==0.10.2.post1
soundfile==0.13.1
torchcrepe==0.0.24
encodec==0.1.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
vocos==0.1.0
julius==0.2.7
lameenc==1.8.1
pyloudnorm==0.1.1
pydub==0.25.1
sox==1.5.0
soxr==1.0.0
nnaudio==0.3.3
av==16.0.1
audioread==3.0.1
resampy==0.4.3
pystoi==0.4.1
torch-stoi==0.2.3
torch-complex==0.4.4
audiobox-aesthetics==0.0.4
torchcodec==0.10.0

# ============ Music / MIDI ============
music21==9.9.1
pretty-midi==0.2.11
midi2audio==0.1.1
mido==1.3.3
pylrc==0.1.2

# ============ Chinese NLP / TTS ============
pypinyin==0.55.0
cn2an==0.5.23
phonemizer==3.3.0
jieba==0.42.1
zhconv==1.4.3
zhon==2.1.1
py3langid==0.3.0
segments==2.3.0

# ============ Model Hubs ============
huggingface-hub==0.35.3
modelscope==1.32.0

# ============ Neural Net Utilities ============
einops==0.8.1
einops-exts==0.0.4
einx==0.3.0
ema-pytorch==0.7.7
rotary-embedding-torch==0.8.9
vector-quantize-pytorch==1.23.2
x-transformers==2.1.2
x-clip==0.14.4
alias-free-torch==0.0.6
loralib==0.1.2
torchdiffeq==0.2.5
torchsde==0.2.6
kornia==0.8.2
clean-fid==0.1.35
clip-anytorch==2.6.0

# ============ Config / CLI ============
hydra-core==1.3.2
omegaconf==2.3.0
fire==0.7.1
argbind==0.3.9
prefigure==0.0.10
gin-config==0.5.0
ml-collections==1.1.0
easydict==1.13
pyyaml==6.0.3

# ============ Web / API (Gradio) ============
gradio

# ============ Monitoring / Logging ============
wandb==0.22.1
loguru==0.7.3

# ============ Data / Science ============

scipy==1.15.3
pandas==2.2.3
scikit-learn==1.7.2
scikit-image==0.25.2
matplotlib==3.10.8
pillow==12.1.1

# ============ Misc Utilities ============
tqdm==4.67.1
regex==2025.9.18
ftfy==6.3.1
unidecode==1.3.8
rapidfuzz==2.13.7
mutagen==1.47.0
numba==0.62.1
onnxruntime==1.23.2
protobuf
qwen-omni-utils==0.0.8
muq==0.1.0
pylance==0.23.2
proces==0.1.7
groovy==0.1.2
randomname==0.2.1
submitit==1.5.4
inflect==7.5.0